fby3.5: cl: Fix mistakenly sending SEL of VPP power control Description: Init the power status of M.2 device by asking SB CPLD register. -> 1. Add plat_power.c and plat_power.h to maintain the power control related function. 2. Move power control related function in plat_isr.c to plat_power.c
Motivation:
Currently we will set power status of M.2 device with default value, which is incorrect. It may cause BIC to mistakenly send SEL to BMC if the device is not inserted.

Test Plan:
1. Build and test pass on YV3.5 system.
pass

2. Check the SEL after setting M.2 device power off.
(1) Before fix
2023 Apr 24 19:02:57 log-util: User cleared all logs
2 slot2 2023-04-24 19:03:10 power-util SERVER_POWER_OFF successful for FRU: 2 DEV: 1U-dev1
2 slot2 2023-04-24 19:03:10 ipmid SEL Entry: FRU: 2, Record: Facebook Unified SEL (0xFB), GeneralInfo: x86/PCIeErr(0x20), Bus 84/Dev 03/Fun 00, 1OU/Num 1,TotalErrID1Cnt: 0x0000, ErrID2: 0x50(Received ERR_COR Message), ErrID1: 0x00(Receiver Error)
2 slot2 2023-04-24 19:03:10 ipmid SEL Entry: FRU: 2, Record: Facebook Unified SEL (0xFB), GeneralInfo: x86/PCIeErr(0x20), Bus 84/Dev 03/Fun 00, 1OU/Num 1,TotalErrID1Cnt: 0x0000, ErrID2: 0x21(Surprise Down Error), ErrID1: 0x53(DPC triggered by uncorrectable error)
2 slot2 2023-04-24 19:03:10 ipmid SEL Entry: FRU: 2, Record: Standard (0x02), Time: 2023-04-24 19:03:10, Sensor: SYSTEM_STATUS (0x10), Event Data: (0B013A) 1OU/Num 0 VPP power control Triggered
2 slot2 2023-04-24 19:03:10 ipmid SEL Entry: FRU: 2, Record: Standard (0x02), Time: 2023-04-24 19:03:10, Sensor: SYSTEM_STATUS (0x10), Event Data: (0B013C) 1OU/Num 1 VPP power control Triggered

(2) After fix
root@bmc-oob:~# power-util slot1 1U-dev1 status
Power status for fru 1 dev 1U-dev1 : ON
root@bmc-oob:~# log-util all --clear
root@bmc-oob:~# power-util slot1 1U-dev1 off
Powering fru 1 dev 1U-dev1 to OFF...
root@bmc-oob:~# log-util all --print
2023 Apr 26 00:59:15 log-util: User cleared all logs
1    slot1    2023-04-26 00:59:18    power-util       SERVER_POWER_OFF successful for FRU: 1 DEV: 1U-dev1
1    slot1    2023-04-26 00:59:18    ipmid            SEL Entry: FRU: 1, Record: Facebook Unified SEL (0xFB), GeneralInfo: x86/PCIeErr(0x20), Bus 84/Dev 03/Fun 00, 1OU/Num 1,TotalErrID1Cnt: 0x0000, ErrID2: 0x50(Received ERR_COR Message), ErrID1: 0x00(Receiver Error)
1    slot1    2023-04-26 00:59:18    ipmid            SEL Entry: FRU: 1, Record: Facebook Unified SEL (0xFB), GeneralInfo: x86/PCIeErr(0x20), Bus 84/Dev 03/Fun 00, 1OU/Num 1,TotalErrID1Cnt: 0x0000, ErrID2: 0x21(Surprise Down Error), ErrID1: 0x53(DPC triggered by uncorrectable error)
1    slot1    2023-04-26 00:59:19    ipmid            SEL Entry: FRU: 1, Record: Standard (0x02), Time: 2023-04-26 00:59:19, Sensor: SYSTEM_STATUS (0x10), Event Data: (0B013C) 1OU/Num 1 VPP power control Triggered
