+++ meta-facebook/yv35-cl/src/platform/plat_power.h
@@ -1,0 +1,28 @@
+/*
+ * Copyright (c) Meta Platforms, Inc. and affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#ifndef PLAT_POWER_H
+#define PLAT_POWER_H
+
+#include <stdint.h>
+#include "ipmi.h"
+
+void init_vpp_power_status();
+uint8_t get_last_vpp_power_status();
+uint8_t get_updated_vpp_power_status();
+int get_set_1ou_m2_power(ipmi_msg *msg, uint8_t device_id, uint8_t option);
+
+#endif
